requirements/base.txt
asgiref==3.3.4
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
Django==3.2.3
django-crispy-forms==1.11.2
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-extensions==3.1.3
django-pandas==0.6.4
django-qsstats-magic==1.1.0
django-rest-authtoken==2.1.3
djangorestframework==3.12.4

iniconfig==1.1.1
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.20.3
packaging==20.9
panda==0.3.1
pandas==1.2.4
Pillow==8.2.0
pluggy==1.0.0.dev0
psycopg2-binary==2.8.6
py==1.10.0
pyparsing==3.0.0b2
pytest==6.2.4
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
